build: use npm prepack instead of prepublishOnly
In NPM, the `prepublishOnly` script is run only before running
`npm publish`. However, this means that people trying to use the git
version of of this project
(i.e. with `npm install ferdikoomen/openapi-typescript-codegen`)
won't get a working version of this project, since there will be
nothing in the `dist/` folder.

Both [Yarn v2][1] and [NPM][2] support a `prepack` script, that is
called whenever `npm pack` is called:

> Runs BEFORE a tarball is packed (on `"npm pack"`, `"npm publish"`,
> and when installing a git dependencies).

[1]: https://yarnpkg.com/advanced/lifecycle-scripts
[2]: https://docs.npmjs.com/cli/v6/using-npm/scripts#life-cycle-scripts